# docs/openapi/bookkin-api.yaml
openapi: 3.1.0
info:
  title: BookKin API
  version: 1.0.0
  description: 家庭 NAS 电子图书馆的浏览、阅读、用户与受控文件事务 API。
servers:
  - url: /api/v1
security:
  - sessionCookie: []
...
paths:
...
  /books:
    get:
      tags: [Catalog]
      operationId: listBooks
      parameters:
        - { name: q, in: query, schema: { type: string } }
        - { name: format, in: query, schema: { $ref: '#/components/schemas/BookFormat' } }
        - { name: sort, in: query, schema: { type: string, enum: [RECENT, TITLE, AUTHOR], default: RECENT } }
        - { name: cursor, in: query, schema: { type: string } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 60 } }
      responses:
        '200': { description: Cursor page, content: { application/json: { schema: { $ref: '#/components/schemas/BookPage' } } } }
...
  /display-books/{id}:
    get:
      tags: [Display catalog]
      security: []
      operationId: getDisplayBook
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses:
        '200': { description: Public book, content: { application/json: { schema: { $ref: '#/components/schemas/DisplayBook' } } } }
    delete:
      tags: [Display catalog]
      operationId: removeDisplayBook
      parameters: [{ $ref: '#/components/parameters/Id' }, { name: revision, in: query, required: true, schema: { type: integer, format: int64, minimum: 0 } }]
      responses:
        '204': { description: Removed from public display catalog }
        '409': { $ref: '#/components/responses/Problem' }
...
  /display-books/{id}/content:
    get:
      tags: [Display catalog]
      security: []
      operationId: streamDisplayBook
      parameters:
        - { $ref: '#/components/parameters/Id' }
        - { name: Range, in: header, schema: { type: string }, example: bytes=0-1048575 }
      responses:
        '200': { description: Public read-only EPUB or PDF, content: { application/octet-stream: { schema: { type: string, format: binary } } } }
        '206': { description: Partial public content, content: { application/octet-stream: { schema: { type: string, format: binary } } } }
...
  /books/{id}/metadata:
    get:
      tags: [Catalog]
      operationId: getBookMetadata
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses:
        '200': { description: Editable metadata, content: { application/json: { schema: { $ref: '#/components/schemas/BookMetadata' } } } }
    patch:
      tags: [Catalog]
      operationId: updateBookMetadata
      parameters: [{ $ref: '#/components/parameters/Id' }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/MetadataRequest' } } } }
      responses:
        '200': { description: Database metadata and optional write-back preview, content: { application/json: { schema: { type: object, required: [metadata], properties: { metadata: { $ref: '#/components/schemas/BookMetadata' }, writeBackPreview: { $ref: '#/components/schemas/FileOperationPreview' } } } } } }
  /books/{id}/cover:
    get:
      tags: [Catalog]
      operationId: getBookCover
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses:
        '200': { description: Current application cover, content: { image/jpeg: { schema: { type: string, format: binary } } } }
    put:
      tags: [Catalog]
      operationId: replaceBookCover
      parameters: [{ $ref: '#/components/parameters/Id' }]
      requestBody: { required: true, content: { image/jpeg: { schema: { type: string, format: binary } }, image/png: { schema: { type: string, format: binary } }, image/webp: { schema: { type: string, format: binary } } } }
      responses:
        '200': { description: Custom cover normalized and activated, content: { application/json: { schema: { $ref: '#/components/schemas/BookMetadata' } } } }
    delete:
      tags: [Catalog]
      operationId: restoreBookCover
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses:
        '200': { description: Custom cover retired and generated or extracted cover restored, content: { application/json: { schema: { $ref: '#/components/schemas/BookMetadata' } } } }
...
  /recycle-bin/{id}/restore:
    post:
      tags: [File operations]
      operationId: restoreRecycleEntry
      parameters: [{ $ref: '#/components/parameters/Id' }, { $ref: '#/components/parameters/IdempotencyKey' }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/FileOperationExecuteRequest' } } } }
      responses:
        '200': { description: Queued restore, content: { application/json: { schema: { $ref: '#/components/schemas/FileOperation' } } } }
  /recycle-bin/{id}:
    delete:
      tags: [File operations]
      operationId: purgeRecycleEntry
      parameters: [{ $ref: '#/components/parameters/Id' }, { $ref: '#/components/parameters/IdempotencyKey' }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/FileOperationExecuteRequest' } } } }
      responses:
        '200': { description: Queued permanent purge, content: { application/json: { schema: { $ref: '#/components/schemas/FileOperation' } } } }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: BOOKKIN_SESSION
  parameters:
    Id: { name: id, in: path, required: true, schema: { type: string, format: uuid } }
    IdempotencyKey: { name: Idempotency-Key, in: header, required: true, schema: { type: string, minLength: 1, maxLength: 160 } }
  responses:
    Problem:
      description: RFC 9457 problem
      content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } }
  schemas:
    Problem:
      type: object
      properties: { type: { type: string, format: uri }, title: { type: string }, status: { type: integer }, detail: { type: string }, code: { type: string } }
...
    BookFormat: { type: string, enum: [EPUB, PDF] }
    BookFileStatus: { type: string, enum: [AVAILABLE, OPERATING, TRASHED, MISSING, DELETED] }
...
    Book:
      type: object
      required: [id, fileId, title, author, format, coverUrl, progress, addedAt, libraryRoot, relativePath, fingerprint, status, tags]
      properties:
        id: { type: string, format: uuid }
        fileId: { type: string, format: uuid }
        title: { type: string }
        author: { type: string }
        series: { type: [string, 'null'] }
        description: { type: [string, 'null'] }
        format: { $ref: '#/components/schemas/BookFormat' }
        coverUrl: { type: string }
        progress: { type: integer, minimum: 0, maximum: 100 }
        wordCount: { type: [string, 'null'] }
        addedAt: { type: string, format: date-time }
        libraryRoot: { type: string }
        relativePath: { type: string }
        fingerprint: { type: string }
        status: { $ref: '#/components/schemas/BookFileStatus' }
        tags: { type: array, items: { type: string } }
    BookPage: { type: object, required: [items], properties: { items: { type: array, items: { $ref: '#/components/schemas/Book' } }, nextCursor: { type: [string, 'null'] } } }
...
    DisplayBook:
      type: object
      required: [id, title, author, coverUrl, available, sortOrder]
      properties:
        id: { type: string, format: uuid }
        title: { type: string }
        author: { type: string }
        description: { type: [string, 'null'] }
        format: { oneOf: [{ $ref: '#/components/schemas/BookFormat' }, { type: 'null' }] }
        coverUrl: { type: string }
        available: { type: boolean }
        sortOrder: { type: integer, format: int64 }
...
    MetadataSource: { type: string, enum: [FILE, FILENAME, OPEN_LIBRARY, GOOGLE_BOOKS, MANUAL] }
...
    BookMetadata:
      type: object
      required: [id, title, authors, translators, tags, sources]
      properties: { id: { type: string, format: uuid }, title: { type: string }, subtitle: { type: [string, 'null'] }, authors: { type: array, items: { type: string } }, translators: { type: array, items: { type: string } }, language: { type: [string, 'null'] }, publisher: { type: [string, 'null'] }, publishedDate: { type: [string, 'null'] }, isbn: { type: [string, 'null'] }, description: { type: [string, 'null'] }, series: { type: [string, 'null'] }, seriesIndex: { type: [number, 'null'] }, tags: { type: array, items: { type: string } }, coverCacheKey: { type: [string, 'null'] }, coverUrl: { type: [string, 'null'] }, sources: { type: object, additionalProperties: { $ref: '#/components/schemas/MetadataSource' } } }
...
    FileOperationType: { type: string, enum: [RENAME, MOVE, TRASH, RESTORE, WRITE_METADATA, PURGE] }
    FileOperationStatus: { type: string, enum: [PLANNED, RUNNING, SUCCEEDED, FAILED, ROLLED_BACK] }
    FileConflict: { type: object, required: [code, message], properties: { code: { type: string }, message: { type: string }, path: { type: [string, 'null'] } } }
...
    FileOperationExecuteRequest:
      type: object
      required: [previewToken, expectedFingerprint]
      properties: { previewToken: { type: string, format: uuid }, expectedFingerprint: { type: string } }
    FileOperationPreview:
      type: object
      required: [previewToken, type, sourcePath, requiredBytes, expectedFingerprint, expiresAt, conflicts, warnings]
      properties: { previewToken: { type: string, format: uuid }, type: { $ref: '#/components/schemas/FileOperationType' }, sourcePath: { type: string }, targetPath: { type: [string, 'null'] }, requiredBytes: { type: integer, format: int64 }, expectedFingerprint: { type: string }, expiresAt: { type: string, format: date-time }, conflicts: { type: array, items: { $ref: '#/components/schemas/FileConflict' } }, warnings: { type: array, items: { type: string } } }
    FileOperation:
      type: object
      required: [id, type, status, sourcePath, createdAt, stage]
      properties: { id: { type: string, format: uuid }, type: { $ref: '#/components/schemas/FileOperationType' }, status: { $ref: '#/components/schemas/FileOperationStatus' }, sourcePath: { type: string }, targetPath: { type: [string, 'null'] }, createdAt: { type: string, format: date-time }, stage: { type: string }, errorCode: { type: [string, 'null'] }, errorDetail: { type: [string, 'null'] } }
...
    MetadataRequest:
      type: object
      properties: { title: { type: string }, subtitle: { type: string }, authors: { type: array, items: { type: string } }, translators: { type: array, items: { type: string } }, language: { type: string }, publisher: { type: string }, publishedDate: { type: string }, isbn: { type: string }, description: { type: string }, series: { type: string }, seriesIndex: { type: number }, tags: { type: array, items: { type: string } }, manualFields: { type: array, items: { type: string } }, writeBack: { type: boolean }, bookFileId: { type: string, format: uuid }, expectedFingerprint: { type: string } }